completed data comparison

+++ LivingLab.Core/Entities/DTO/EnergyUsageDTOs/EnergyComparisonGraphDTO.cs
@@ -4,7 +4,7 @@ public class EnergyComparisonGraphDTO
 {
     public string? LabLocation { get; set; }
 
-    public int TotalEnergyUsage { get; set; }
+    public double TotalEnergyUsage { get; set; }
 
     public double EnergyUsageIntensity { get; set; }
